% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Getting data from oTree server}
\usage{
get_data(website)
}
\arguments{
\item{website}{the http url to the website with the data}
}
\description{
Getting data from oTree server
}
